convertTypes
