+++ libraries/transformParamsUtils/convertTypes.js
@@ -1,0 +1,36 @@
+import {isFn} from '../../src/utils.js';
+
+/**
+ * Try to convert a value to a type.
+ * If it can't be done, the value will be returned.
+ *
+ * @param {string} typeToConvert The target type. e.g. "string", "number", etc.
+ * @param {*} value The value to be converted into typeToConvert.
+ */
+function tryConvertType(typeToConvert, value) {
+  if (typeToConvert === 'string') {
+    return value && value.toString();
+  } else if (typeToConvert === 'number') {
+    return Number(value);
+  } else {
+    return value;
+  }
+}
+
+export function convertTypes(types, params) {
+  Object.keys(types).forEach(key => {
+    if (params[key]) {
+      if (isFn(types[key])) {
+        params[key] = types[key](params[key]);
+      } else {
+        params[key] = tryConvertType(types[key], params[key]);
+      }
+
+      // don't send invalid values
+      if (isNaN(params[key])) {
+        delete params.key;
+      }
+    }
+  });
+  return params;
+}
