+++ modules/adrelevantisBidAdapter.js
@@ -1,7 +1,6 @@
 import {Renderer} from '../src/Renderer.js';
 import {
   chunk,
-  convertTypes,
   createTrackPixelHtml,
   deepAccess,
   deepClone,
@@ -22,6 +21,7 @@ import {INSTREAM, OUTSTREAM} from '../src/video.js';
 import { convertOrtbRequestToProprietaryNative } from '../src/native.js';
 import {getANKeywordParam, transformBidderParamKeywords} from '../libraries/appnexusUtils/anKeywords.js';
 import {convertCamelToUnderscore} from '../libraries/appnexusUtils/anUtils.js';
+import {convertTypes} from '../libraries/transformParamsUtils/convertTypes.js';
 
 const BIDDER_CODE = 'adrelevantis';
 const URL = 'https://ssp.adrelevantis.com/prebid';
