+++ modules/connectadBidAdapter.js
@@ -1,8 +1,9 @@
-import { deepSetValue, convertTypes, logWarn } from '../src/utils.js';
+import { deepSetValue, logWarn } from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { BANNER } from '../src/mediaTypes.js'
 import {config} from '../src/config.js';
 import {tryAppendQueryString} from '../libraries/urlUtils/urlUtils.js';
+import {convertTypes} from '../libraries/transformParamsUtils/convertTypes.js';
 const BIDDER_CODE = 'connectad';
 const BIDDER_CODE_ALIAS = 'connectadrealtime';
 const ENDPOINT_URL = 'https://i.connectad.io/api/v2';
