+++ modules/openxBidAdapter.js
@@ -4,6 +4,7 @@ import * as utils from '../src/utils.js';
 import {mergeDeep} from '../src/utils.js';
 import {BANNER, VIDEO} from '../src/mediaTypes.js';
 import {ortbConverter} from '../libraries/ortbConverter/converter.js';
+import {convertTypes} from '../libraries/transformParamsUtils/convertTypes.js';
 
 const bidderConfig = 'hb_pb_ortb';
 const bidderVersion = '2.0';
@@ -149,7 +150,7 @@ const converter = ortbConverter({
 });
 
 function transformBidParams(params, isOpenRtb) {
-  return utils.convertTypes({
+  return convertTypes({
     'unit': 'string',
     'customFloor': 'number'
   }, params);
