+++ modules/rubiconBidAdapter.js
@@ -7,7 +7,6 @@ import { find } from '../src/polyfill.js';
 import { getGlobal } from '../src/prebidGlobal.js';
 import { Renderer } from '../src/Renderer.js';
 import {
-  convertTypes,
   deepAccess,
   deepSetValue,
   formatQS,
@@ -21,6 +20,7 @@ import {
   parseSizesInput, _each
 } from '../src/utils.js';
 import {getAllOrtbKeywords} from '../libraries/keywords/keywords.js';
+import {convertTypes} from '../libraries/transformParamsUtils/convertTypes.js';
 
 const DEFAULT_INTEGRATION = 'pbjs_lite';
 const DEFAULT_PBS_INTEGRATION = 'pbjs';
