+++ modules/trafficgateBidAdapter.js
@@ -1,7 +1,8 @@
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER, VIDEO} from '../src/mediaTypes.js';
 import {ortbConverter} from '../libraries/ortbConverter/converter.js';
-import {deepAccess, mergeDeep, convertTypes} from '../src/utils.js';
+import {deepAccess, mergeDeep} from '../src/utils.js';
+import {convertTypes} from '../libraries/transformParamsUtils/convertTypes.js';
 
 const BIDDER_CODE = 'trafficgate';
 const URL = 'https://[HOST].bc-plugin.com/prebidjs'
